+++ MIMS.Entity/Model/PHA_StockDetail.cs
@@ -36,27 +36,27 @@ namespace MIMS.Entity.Model
         /// <summary>
         /// InWarehousePrice
         /// </summary>		
-        public decimal InWarehousePrice { get; set; }
+        public double InWarehousePrice { get; set; }
         /// <summary>
         /// WholesalePrice
         /// </summary>		
-        public decimal WholesalePrice { get; set; }
+        public double WholesalePrice { get; set; }
         /// <summary>
         /// RetailPrice
         /// </summary>		
-        public decimal RetailPrice { get; set; }
+        public double RetailPrice { get; set; }
         /// <summary>
         /// RetailSum
         /// </summary>		
-        public decimal RetailSum { get; set; }
+        public double RetailSum { get; set; }
         /// <summary>
         /// WholesaleSum
         /// </summary>		
-        public decimal WholesaleSum { get; set; }
+        public double WholesaleSum { get; set; }
         /// <summary>
         /// InWarehouseSum
         /// </summary>		
-        public decimal InWarehouseSum { get; set; }
+        public double InWarehouseSum { get; set; }
 
     }
 }
