+++ MIMS.Entity/Model/PSS_ExWarehouseDetail.cs
@@ -32,15 +32,15 @@ namespace MIMS.Entity.Model
         /// <summary>
         /// InWarehousePrice
         /// </summary>		
-        public decimal InWarehousePrice { get; set; }
+        public double InWarehousePrice { get; set; }
         /// <summary>
         /// WholesalePrice
         /// </summary>		
-        public decimal WholesalePrice { get; set; }
+        public double WholesalePrice { get; set; }
         /// <summary>
         /// RetailPrice
         /// </summary>		
-        public decimal RetailPrice { get; set; }
+        public double RetailPrice { get; set; }
         /// <summary>
         /// ApplyNum
         /// </summary>
